refactor: split app route state helpers
